finalized BigDataService
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C32363

+++ Vanrise/Code/Vanrise.Common.Business/Bigdata/BigDataRuntimeService.cs
@@ -18,43 +18,67 @@ namespace Vanrise.Common.Business
 
         int _wcfPortRangeStart;
         int _wcfPortRangeEnd;
+        int _wcfServiceHostingRetries;
 
-        protected override void OnStarted()
+        TimeSpan _deleteTimedOutServicesInterval;
+        DateTime _deleteTimedOutServicesLastTime;
+
+        public BigDataRuntimeService()
         {
-            if(!int.TryParse(ConfigurationManager.AppSettings["WCFPortRangeStart"], out _wcfPortRangeStart))
+            if (!int.TryParse(ConfigurationManager.AppSettings["WCFPortRangeStart"], out _wcfPortRangeStart))
                 _wcfPortRangeStart = 40000;
             if (!int.TryParse(ConfigurationManager.AppSettings["WCFPortRangeEnd"], out _wcfPortRangeEnd))
                 _wcfPortRangeEnd = 50000;
+            if (!int.TryParse(ConfigurationManager.AppSettings["WCFServiceHostingRetries"], out _wcfServiceHostingRetries))
+                _wcfServiceHostingRetries = 100;
+            if (!TimeSpan.TryParse(ConfigurationManager.AppSettings["BigDataCache_DeleteTimedOutServicesInterval"], out _deleteTimedOutServicesInterval))
+                _deleteTimedOutServicesInterval = TimeSpan.FromMinutes(2);
+        }
+
+
+        protected override void OnStarted()
+        {
             BigDataManager.Instance._isBigDataHost = true;
-            _serviceHost = new ServiceHost(typeof(BigDataWCFService));
-            string machineName = Environment.MachineName;
             var random = new Random();
-            string portNumber = random.Next(_wcfPortRangeStart, _wcfPortRangeEnd).ToString();
-            string serviceUrl = String.Format("net.tcp://{0}:{1}/BigDataWCFService", machineName, portNumber);
-            var endPoint = AddTCPEndPoint(_serviceHost, serviceUrl);
-            _serviceHost.Opening += serviceHost_Opening;
-            _serviceHost.Opened += new EventHandler(serviceHost_Opened);
-            _serviceHost.Closing += serviceHost_Closing;
-            _serviceHost.Closed += serviceHost_Closed;
-            for (int i = 0; i < 10; i++)
+
+            string serviceUrl = null;
+            for (int i = 0; i < _wcfServiceHostingRetries; i++)
             {
+                string portNumber = random.Next(_wcfPortRangeStart, _wcfPortRangeEnd).ToString();
+                serviceUrl = String.Format("net.tcp://{0}:{1}/BigDataWCFService", Environment.MachineName, portNumber);
+
                 try
                 {
-                    portNumber = random.Next(_wcfPortRangeStart, _wcfPortRangeEnd).ToString();
-                    serviceUrl = String.Format("net.tcp://{0}:{1}/BigDataWCFService", machineName, portNumber);
-                    endPoint.Address = new EndpointAddress(serviceUrl);
-                    _serviceHost.Open();
+                    CreateServiceHost(serviceUrl);
+                    LoggerFactory.GetLogger().WriteInformation("Service URL registered successfully '{0}'", serviceUrl);
                     break;
                 }
                 catch
                 {
+                    _serviceHost.Opening -= serviceHost_Opening;
+                    _serviceHost.Opened -= serviceHost_Opened;
+                    _serviceHost.Closing -= serviceHost_Closing;
+                    _serviceHost.Closed -= serviceHost_Closed;
                     LoggerFactory.GetLogger().WriteWarning("Could not register '{0}'", serviceUrl);
-                    throw ;
+                    if (i == (_wcfServiceHostingRetries - 1))//last iteration
+                        throw;
                 }
             }
             if (!_dataManager.Insert(serviceUrl, Vanrise.Runtime.RunningProcessManager.CurrentProcess.ProcessId, out _bigDataServiceId))
                 throw new Exception("Could not insert BigDataService into database");
             base.OnStarted();
+        }
+
+        private void CreateServiceHost(string serviceUrl)
+        {
+            _serviceHost = new ServiceHost(typeof(BigDataWCFService));
+            var endPoint = AddTCPEndPoint(_serviceHost, serviceUrl);
+            _serviceHost.Opening += serviceHost_Opening;
+            _serviceHost.Opened += serviceHost_Opened;
+            _serviceHost.Closing += serviceHost_Closing;
+            _serviceHost.Closed += serviceHost_Closed;
+
+            _serviceHost.Open();
         }
 
         protected override void OnStopped()
@@ -68,11 +92,11 @@ namespace Vanrise.Common.Business
 
         private ServiceEndpoint AddTCPEndPoint(ServiceHost serviceHost, string serviceUrl)
         {
-            LoggerFactory.GetLogger().WriteInformation("Adding NetTcp End Point to Business Process WCF Service (BPService)...");
+            //LoggerFactory.GetLogger().WriteInformation("Adding NetTcp End Point to Business Process WCF Service (BPService)...");
             NetTcpBinding binding = new NetTcpBinding(SecurityMode.None);
             binding.MaxReceivedMessageSize = int.MaxValue;
             var endPoint = serviceHost.AddServiceEndpoint(typeof(IBigDataWCFService), binding, serviceUrl);
-            LoggerFactory.GetLogger().WriteInformation("NetTcp End Point to Business Process WCF Service (BPService) added");
+            //LoggerFactory.GetLogger().WriteInformation("NetTcp End Point to Business Process WCF Service (BPService) added");
             return endPoint;
         }
 
@@ -103,9 +127,13 @@ namespace Vanrise.Common.Business
                 _dataManager.Update(_bigDataServiceId, BigDataManager.Instance._totalRecordsCount, BigDataManager.Instance.CachedObjectIds);
                 BigDataManager.Instance._isCachedDataChanged = false;
             }
-            Vanrise.Runtime.RunningProcessManager runningProcessManager = new Runtime.RunningProcessManager();
-            IEnumerable<int> runningProcessIds = runningProcessManager.GetCachedRunningProcesses().MapRecords(itm => itm.ProcessId);
-            _dataManager.DeleteTimedOutServices(runningProcessIds);
+            if ((DateTime.Now - _deleteTimedOutServicesLastTime) > _deleteTimedOutServicesInterval)
+            {
+                Vanrise.Runtime.RunningProcessManager runningProcessManager = new Runtime.RunningProcessManager();
+                IEnumerable<int> runningProcessIds = runningProcessManager.GetCachedRunningProcesses().MapRecords(itm => itm.ProcessId);
+                _dataManager.DeleteTimedOutServices(runningProcessIds);
+                _deleteTimedOutServicesLastTime = DateTime.Now;
+            }
         }
     }
 }
